+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * doc comments
 
+### Changed
+* I reorganized the dependencies in `Cargo.toml`.
+
 
 ## [0.1.17] (2026-04-16)
 ### Changed